Use a common saga id generator implementation
