+++ src/NServiceBus.Persistence.CosmosDB/Saga/CosmosDBSagaIdGenerator.cs
@@ -1,0 +1,32 @@
+﻿namespace NServiceBus.Persistence.CosmosDB
+{
+    using System;
+    using System.Security.Cryptography;
+    using System.Text;
+    using Newtonsoft.Json;
+
+    static class CosmosDBSagaIdGenerator
+    {
+        public static Guid Generate(Type sagaEntityType, string correlationPropertyName, object correlationPropertyValue) => Generate(sagaEntityType.FullName, correlationPropertyName, correlationPropertyValue);
+
+        public static Guid Generate(string sagaEntityTypeFullName, string correlationPropertyName, object correlationPropertyValue)
+        {
+            // assumes single correlated sagas since v6 doesn't allow more than one corr prop
+            // will still have to use a GUID since moving to a string id will have to wait since its a breaking change
+            var serializedPropertyValue = JsonConvert.SerializeObject(correlationPropertyValue);
+            return DeterministicGuid($"{sagaEntityTypeFullName}_{correlationPropertyName}_{serializedPropertyValue}");
+        }
+
+        static Guid DeterministicGuid(string src)
+        {
+            var stringBytes = Encoding.UTF8.GetBytes(src);
+
+            using (var sha1CryptoServiceProvider = new SHA1CryptoServiceProvider())
+            {
+                var hashedBytes = sha1CryptoServiceProvider.ComputeHash(stringBytes);
+                Array.Resize(ref hashedBytes, 16);
+                return new Guid(hashedBytes);
+            }
+        }
+    }
+}
